backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2==2.9.10
python-jose==3.3.0
django-celery-beat==2.5.0
daphne==4.0.0
django-environ==0.11.2
drf-yasg==1.21.7
celery==5.3.6
redis==5.0.1
gunicorn==21.2.0
rest_framework-simplejwt
djangorestframework-datatables
python-jwt==4.1.0
django-redis==5.4.0
channels-redis==4.2.0
environ
validators==0.34.0
XlsxWriter==3.2.2
django-filter==24.3
requests==2.32.3
python-json-logger==2.0.7
djangorestframework-simplejwt==5.4.0
pysocks==1.7.1
dnspython==2.7.0
python-whois==0.9.5
playwrightcapture==1.27.8
tldextract==5.1.3
vt-py==0.20.0
psycopg2-binary==2.9.10